Extract report functionality out of main
